+++ api/src/main/java/org/allaymc/api/blockentity/interfaces/BlockEntityShelf.java
@@ -1,10 +1,10 @@
 package org.allaymc.api.blockentity.interfaces;
 
 import org.allaymc.api.blockentity.BlockEntity;
-import org.allaymc.api.blockentity.component.BlockEntityShelfBaseComponent;
+import org.allaymc.api.blockentity.component.BlockEntityContainerHolderComponent;
 
 /**
  * @author daoge_cmd
  */
-public interface BlockEntityShelf extends BlockEntity, BlockEntityShelfBaseComponent {
+public interface BlockEntityShelf extends BlockEntity, BlockEntityContainerHolderComponent {
 }