+++ server/src/main/java/org/allaymc/server/blockentity/impl/BlockEntityShelfImpl.java
@@ -2,7 +2,7 @@ package org.allaymc.server.blockentity.impl;
 
 import lombok.experimental.Delegate;
 import org.allaymc.api.blockentity.BlockEntityInitInfo;
-import org.allaymc.api.blockentity.component.BlockEntityShelfBaseComponent;
+import org.allaymc.api.blockentity.component.BlockEntityContainerHolderComponent;
 import org.allaymc.api.blockentity.interfaces.BlockEntityShelf;
 import org.allaymc.api.component.Component;
 import org.allaymc.server.component.ComponentProvider;
@@ -14,7 +14,7 @@ import java.util.List;
  */
 public class BlockEntityShelfImpl extends BlockEntityImpl implements BlockEntityShelf {
     @Delegate
-    private BlockEntityShelfBaseComponent shelfBaseComponent;
+    private BlockEntityContainerHolderComponent containerHolderComponent;
 
     public BlockEntityShelfImpl(BlockEntityInitInfo initInfo, List<ComponentProvider<? extends Component>> componentProviders) {
         super(initInfo, componentProviders);